+++ components/ui/avatar.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as AvatarPrimitive from "@radix-ui/react-avatar";
 
-import {cn} from "~/utils/className";
+import {cn} from "~/utils/cn";
 
 const Avatar = React.forwardRef<
   React.ElementRef<typeof AvatarPrimitive.Root>,
