+++ components/ui/button.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import {Slot} from "@radix-ui/react-slot";
 import {cva, type VariantProps} from "class-variance-authority";
 
-import {cn} from "~/utils/className";
+import {cn} from "~/utils/cn";
 
 const buttonVariants = cva(
   `font-poppins font-medium shrink-0
