+++ components/ui/calendar.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import {DayPicker} from "react-day-picker";
 import {ChevronLeftIcon, ChevronRightIcon} from "lucide-react";
 
-import {cn} from "~/utils/className";
+import {cn} from "~/utils/cn";
 
 const Calendar = ({classNames, ...props}: React.ComponentProps<typeof DayPicker>) => {
   return (
