+++ components/ui/dialog.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
 import {XIcon} from "lucide-react";
 
-import {cn} from "~/utils/className";
+import {cn} from "~/utils/cn";
 
 const Dialog = DialogPrimitive.Root;
