+++ components/ui/dropdown-menu.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
 
-import {cn} from "~/utils/className";
+import {cn} from "~/utils/cn";
 
 const DropdownMenu = DropdownMenuPrimitive.Root;
