+++ components/ui/popover.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as PopoverPrimitive from "@radix-ui/react-popover";
 
-import {cn} from "~/utils/className";
+import {cn} from "~/utils/cn";
 
 const Popover = PopoverPrimitive.Root;
