+++ components/ui/separator.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import {cn} from "~/utils/className";
+import {cn} from "~/utils/cn";
 
 type SeparatorProps = {
   orientation?: "horizontal" | "vertical";
