+++ components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import {cn} from "~/utils/className";
+import {cn} from "~/utils/cn";
 
 function Skeleton({className, ...props}: React.HTMLAttributes<HTMLDivElement>) {
   return <div className={cn("animate-pulse rounded-md bg-muted", className)} {...props} />;
