+++ components/ui/textarea.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import {cn} from "~/utils/className";
+import {cn} from "~/utils/cn";
 
 type TextareaProps = React.TextareaHTMLAttributes<HTMLTextAreaElement>;
